requirements.txt
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
boto3==1.35.5
botocore==1.35.5
certifi==2024.7.4
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.15.4
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
jiter==0.5.0
jmespath==1.0.1
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
